Change bipartite site index value in case of partial initialization.
Unlike short site index, bipartite site index has additional function.
It points to the current site in the value array. That is why same
index value could mean two different site states: either partially
initialized or completely initialized and processed. It was additional
argument in the countup_res function to distingush the states.
Currently the index has negative value in the case of incomplete
initialization and that additional argument was removed.
cbclib.counts